feat: add login page and onboarding functionality
Adds a new login page and integrates onboarding functionality into the app.

The onboarding process uses a card swiper to
 display welcome screens and includes a blur button to navigate to the login page.
The app checks if the user has seen the onboarding screens before and redirects them to the login page if they have.
Shared preferences are used to store the onboarding status.
Additionally, the Podfile is added for iOS dependencies.