fix: prevent back button navigation from products page
Prevents users from navigating back to the home page using the back button when they are on the products page.

This change ensures that users return to the app page when pressing the back button on the products page, maintaining the desired navigation flow.
Additionally, a buildWhen condition has been added to the home page BlocConsumer to prevent unnecessary rebuilds when the state changes but the categories remain the same.
This optimization improves the
 performance of the home page by reducing redundant rebuilds.
Lastly, the internet permission has been added to the manifest file to allow network access for the application.