feat: improve cart page UI and add functionality
Adds a new UI for the cart page with a checkout button and a total price display.
Implements functionality
 to update and remove items from the cart.
Includes a dismissible widget to remove items from the cart by swiping.
Displays an empty cart message when there are no items in the cart.
Adds a placeholder widget at the end of the cart list to prevent the checkout button from overlapping the last item.

Updates the cart database to include a `createdAt` field.
Adds a new `IncreaseDecreaseButton` widget to the product detail page and cart page for quantity selection.
Fixes an issue where adding a product to the cart would not update the cart page.
Updates the dependencies in `injection.config.dart`.