feat: add notification page
Adds a new notification page to the app.

The notification page displays a list of notifications for the user.
Users can tap
 on a notification to mark it as read.
Users can swipe left on a notification to delete it.
If there are no notifications, a message is displayed.
The page includes an app bar with a title and a more options button.
The page uses the `NotificationBloc` to manage the state of the
 notifications.